Added name-only State constructors (Extensions and General)
Rather than try to interpret the parameters, I instead copied the
default constructor and overloaded it to take a name only. I then used
super to create the extension constructors with the names End and Start.
One test case examines that the new overload of the name-only
constructor works for the generic and extension constructors.