Little, yet major bug hotfixed.
The DrawerView Read method uses a boolean 'found' to exit scopes quickly
and move on to the next element that needs to be connected. Each element
that is being connected is a transition. It was originally moving on to
the next state after finding one incoming transition, which is wrong.
Now, it moves on to the next transition.