Nested states and nested start state can be created!
Nested start states are added by right clicking the desired parent and
selecting the add nested start option. Regular nested states can be
built from the rounded rectangle button with the 'I' on it. Updating the
parent state's name does NOT update the nested start yet. This is a bug
that needs to be fixed.
In addition, Simulation now works. Simulation is run on a separate
thread. It throws NullPointerExceptions during execution, but it does
not cause execution to end. The null pointer is the cached drawing
bounds, and this may simply be a race condition.